Visual feedback when cupcake is smashed
Made changes in CSS so that when the player clicks on the cupcake,
the cupcake changes into a speech bubble saying "SMASH".
- Added new CSS selector for "smashed" class.
- This selector defines the image shown when the cupcake is being
  clicked on.

This gives the player extra feedback when they have clicked on the
cupcake.

In addition, changed the JS so that the "SMASH" image disappears
again after 0.5 seconds. This works by using a timeout function that
removes the "smashed" CSS class from the classlist.