Fixed cupcake coming back up after being smashed
Previously in CSS, the cupcake element contained properties for
setting the image, position and transition of the cupcake. So after
JavaScript removes the "smashed" class name from the cupcake element,
the class name goes back to "cupcake", and the game shows the
cupcake again even though it was clicked on. A player could keep
clicking on that same cupcake and it would keep reappearing, and keep
adding to the score.

In CSS:
- I created additional class names for each position a cupcake
could be in throughout the game (up, down, smash);
-The main cupcake class now only contains the properties of the
cupcake that don't change throughout the game;
- The changing properties are added to the corresponding new class
names (cupcake up, cupcake down, cupcake smash);
- Removed "up" class name from the hole element.

Updated JavaScript code so that new class names are added to or
removed from the cupcake class at the right moment in the game.